% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predef_lab.R
\name{init_standard_library}
\alias{init_standard_library}
\title{Initialize the Standard Phrases Library}
\usage{
init_standard_library(con)
}
\arguments{
\item{con}{DBIConnection. An active connection to the SQLite database.}
}
\value{
Invisible NULL.
}
\description{
Creates a standardized repository of reusable text clauses (Standard Clauses)
in the database. This allows common pharmaceutical formulations to be managed
centrally and referenced across multiple documents.
}
