% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{search_content}
\alias{search_content}
\title{Full-Text Content Search}
\usage{
search_content(con, term)
}
\arguments{
\item{con}{DBIConnection. Active SQLite database connection.}

\item{term}{Character. The search term or keyword.}
}
\value{
A data.frame containing instance_id, label, and the matching content.
}
\description{
Performs a keyword search within the 'content' attribute of all instances
stored in the database.
}
